add ETL mode
